Migrate to NS 8
https://blog.nativescript.org/nativescript-8-announcement/index.html#updating-to-nativescript-8
