+++ .migration_backup/_backup.json
@@ -1,0 +1,12 @@
+{
+	"name": "migration",
+	"paths": [
+		"hooks",
+		"webpack.config.js",
+		"package.json",
+		"package-lock.json",
+		"app/package.json",
+		"package.json",
+		"app/package.json"
+	]
+}
