+++ .migration_backup/app/package.json
@@ -1,0 +1,7 @@
+{
+  "main": "app.js",
+  "android": {
+    "v8Flags": "--expose_gc",
+    "markingMode": "none"
+  }
+}
